refactor: Delegate ChallengeResponseActivity business code to view model
Delegate ChallengeResponseActivity business logic code to a view model.
This consists of introducing Hilt, Molecule and Compose Runtime as
dependencies and heavily changing the architecture of the UI layer to a
Unidirectional Data Flow. The view model emits UI states on which the
activity listens and reacts upon. The activity in turn emits user
interaction events to the view model which will change the UI state.

compileSdkVersion is bumped to 33 as it's a requirement for
Compose 1.3.2

ConnectionManager, ConnectionMethods and the receivers interfaces were
made public.

SlotPreferenceManager is now an injected singleton.

There is no longer an activity recreation when unplugging the YubiKey.
The initialization of connection in ConnectionManager is now exposed as
initConnection method in order to reuse the class.

Some UI changes have been made to better match the states like only
displaying the loading indicator when communicating with the YubiKey and
also displaying "Please wait." to further indicate communication.